Move RegistryClient from maven to common
