+++ maven/lib/dependabot/maven/file_parser/repositories_finder.rb
@@ -4,6 +4,7 @@ require "nokogiri"
 
 require "dependabot/dependency_file"
 require "dependabot/maven/file_parser"
+require "dependabot/registry_client"
 require "dependabot/errors"
 
 # For documentation, see:
@@ -109,7 +110,7 @@ module Dependabot
             url = remote_pom_url(group_id, artifact_id, version, base_url)
 
             @maven_responses ||= {}
-            @maven_responses[url] ||= RegistryClient.get(
+            @maven_responses[url] ||= Dependabot::RegistryClient.get(
               url: url,
               # We attempt to find dependencies in private repos before failing over to the CENTRAL_REPO_URL,
               # but this can burn a lot of a job's time against slow servers due to our `read_timeout` being 20 seconds.
